+++ tools/helper.py
@@ -1,4 +1,6 @@
-from colorama import Fore, Back, Style
+from colorama import Fore, Style
+import networkx as nx
+import random
 
 
 def print_progress(current, total, bar_length=65):
@@ -19,3 +21,91 @@ def print_progress(current, total, bar_length=65):
 
     # Print the progress bar with color, overwriting the previous line
     print(f"\r{color}{progress_bar}{Style.RESET_ALL}", end="")
+
+
+def create_ppi_network(fly_interactome, fly_GO_term):
+    print("Initializing network")
+    i = 1
+    total_progress = len(fly_interactome) + len(fly_GO_term)
+    G = nx.Graph()
+    protein_protein_edge = 0
+    protein_go_edge = 0
+    protein_node = 0
+    go_node = 0
+    protein_list = []
+    go_term_list = []
+
+    # go through fly interactome, add a new node if it doesnt exists already, then add their physical interactions as edges
+    for line in fly_interactome:
+        if not G.has_node(line[2]):
+            G.add_node(line[2], name=line[0], type="protein")
+            protein_list.append({"id": line[2], "name": line[0]})
+            protein_node += 1
+
+        if not G.has_node(line[3]):
+            G.add_node(line[3], name=line[1], type="protein")
+            protein_list.append({"id": line[3], "name": line[1]})
+            protein_node += 1
+
+        G.add_edge(line[2], line[3], type="protein_protein")
+        protein_protein_edge += 1
+        print_progress(i, total_progress)
+        i += 1
+
+    # Proteins annotated with a GO term have an edge to a GO term node
+    for line in fly_GO_term:
+        if not G.has_node(line[1]):
+            G.add_node(line[1], type="go_term")
+            go_term_list.append(line[1])
+            go_node += 1
+
+        if not G.has_node(line[0]):
+            G.add_node(line[0], name=line[0], type="protein")
+            protein_list.append({"id": line[0], "name": line[0]})
+            protein_node += 1
+
+        G.add_edge(line[1], line[0], type="protein_go_term")
+        protein_go_edge += 1
+        print_progress(i, total_progress)
+        i += 1
+
+    print("")
+    print("")
+    print("network summary")
+
+    print("protein-protein edge count: ", protein_protein_edge)
+    print("protein-go edge count: ", protein_go_edge)
+    print("protein node count: ", protein_node)
+    print("go node count: ", go_node)
+    print("total edge count: ", len(G.edges()))
+    print("total node count: ", len(G.nodes()))
+
+    return G, protein_list, go_term_list
+
+
+def read_specific_columns(file_path, columns, delimit):
+    try:
+        with open(file_path, "r") as file:
+            next(file)
+            data = []
+            for line in file:
+                parts = line.strip().split(delimit)
+                selected_columns = []
+                for col in columns:
+                    selected_columns.append(parts[col].replace('"', ""))
+                data.append(selected_columns)
+            return data
+    except FileNotFoundError:
+        print(f"Error: File '{file_path}' not found.")
+        return None
+    except Exception as e:
+        print(f"An error occurred: {e}")
+        return None
+
+
+def generate_random_colors(num_colors):
+    colors = []
+    for _ in range(num_colors):
+        color = (random.random(), random.random(), random.random())
+        colors.append(color)
+    return colors
