+++ tools/workflow.py
@@ -1,0 +1,53 @@
+from sklearn.metrics import roc_curve, auc, precision_recall_curve
+
+
+def run_workflow(
+    algorithm_classes, positive_dataset, negative_dataset, G, output_data_path
+):
+    results = {}
+    for algorithm_name, algorithm_class in algorithm_classes.items():
+        current = run_algorithm(
+            algorithm_class, positive_dataset, negative_dataset, G, output_data_path
+        )
+        current = run_metrics(current)
+        results[algorithm_name] = current
+    return results
+
+
+def run_algorithm(
+    algorithm_class,
+    positive_dataset,
+    negative_dataset,
+    G,
+    output_data_path,
+):
+    # Create an instance of the algorithm class
+    algorithm = algorithm_class()
+
+    # Predict using the algorithm
+    algorithm.predict(positive_dataset, negative_dataset, G, output_data_path)
+
+    # Access y_true and y_score attributes for evaluation
+    y_true = algorithm.y_true
+    y_score = algorithm.y_score
+
+    results = {"y_true": y_true, "y_score": y_score}
+
+    return results
+
+
+def run_metrics(current):
+    # Compute ROC curve and ROC area for a classifier
+    current["fpr"], current["tpr"], current["thresholds"] = roc_curve(
+        current["y_true"], current["y_score"]
+    )
+    current["roc_auc"] = auc(current["fpr"], current["tpr"])
+
+    # Compute precision-recall curve and area under the curve for a classifier
+    current["precision"], current["recall"], _ = precision_recall_curve(
+        current["y_true"], current["y_score"]
+    )
+    current["pr_auc"] = auc(current["recall"], current["precision"])
+
+    return current
+
